Restructure Affiliates page into All / Pending Payouts tabs
Page heading becomes 'Affiliates' and gains two tabs:
- All Affiliates: new browse UI with search, status filter, paginated
  table of every affiliate from GoAffPro
- Pending Payouts: existing payout/transfer flow, extracted into its
  own component for clarity

Tab content components moved into _components/. The selected-affiliate
state in the page wraps the list for a follow-up detail dialog.
